# docs/swagger.yaml
basePath: /reth
definitions:
...
  info_handlers.ReqUnstakingPlanExist:
    properties:
      stakerAddress:
        description: hex string
        type: string
    type: object
  info_handlers.ReqUploadUnstakingPlan:
    properties:
      amount:
        type: string
      stakerAddress:
        description: hex string
        type: string
    type: object
  info_handlers.ReqWithdrawRemainingTime:
    properties:
      stakerAddress:
        description: hex string
        type: string
    type: object
...
  info_handlers.RspUnstakingPlanExist:
    properties:
      exist:
        type: boolean
    type: object
  info_handlers.RspWithdrawRemainingTime:
    properties:
      remainingSeconds:
        description: staked waiting actived
        type: integer
    type: object
...
  utils.Rsp:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8083
info:
  contact:
    email: tpkeeper.me@gmail.com
    name: tpkeeper
  description: reth api document.
  title: reth API
  version: "1.0"
paths:
...
  /v1/staker/unstakingPlanExist:
    post:
      consumes:
      - application/json
      description: staker unstaking plan exit
      parameters:
      - description: unstaking plan exist
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/info_handlers.ReqUnstakingPlanExist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Rsp'
            - properties:
                data:
                  $ref: '#/definitions/info_handlers.RspUnstakingPlanExist'
              type: object
      summary: unstaking plan exit
      tags:
      - v1
  /v1/staker/uploadUnstakingPlan:
    post:
      consumes:
      - application/json
      description: staker unstaking plan
      parameters:
      - description: unstaking plan
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/info_handlers.ReqUploadUnstakingPlan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Rsp'
      summary: unstaking plan
      tags:
      - v1
  /v1/staker/withdrawRemainingTime:
    post:
      consumes:
      - application/json
      description: staker withdraw remaining time
      parameters:
      - description: staker address
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/info_handlers.ReqWithdrawRemainingTime'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Rsp'
            - properties:
                data:
                  $ref: '#/definitions/info_handlers.RspWithdrawRemainingTime'
              type: object
      summary: staker withdraw remaining time
      tags:
      - v1
...
swagger: "2.0"